elx-package-metadata: do not merge old data
We have no old data that could be merged but we do have many implementations
to do it.  Instead of keeping all or some of them in the code and make fixes
with no benefit remove them all and add an optimized solution once we actually
have metadata that should be merged.