elx-elip-files[-1|git]: cleanup, only ignore starting with a dot
These functions previously were more flexible but it turned out that
extending the exclude rules likely results in false positives.  So
elx-elisp-files-exclude only matches dots now.  Also the regexp used
to detect elisp files has been moved to elx-elisp-files-exclude.
Both this variables are constants now as they are not intended to be
changed.  SOURCE argument has been replaced with ALL, so it is not
possible anymore to specify a different exclude regexp (aside of
dynamic binding).