split into two package; bump to version 0.6
  elx   limited to libraries
  xpkg  aware of packages inside git repositories

Function elx-required-packages was lost in the process.  Some more
functionality is now only available when working with packages that are
stored in git repositories.  While it might be desirable to have that
functionality for packages not stores in git repos I simply do not need
that and am happy to remove some unnecessary (to me) complexity.