remove elx-version
Instead just use (lm-header "version").  When the version to be
extracted is sane then the returned value is identical to what
`elx-version' would have returned.  Otherwise you could use:

  (--when-let (lm-header "version")
    (or (vcomp-prefixed-version-p it)
        (vcomp-normalize it)))

This will return the same version, except for libraries by Drew Adams.
For his libraries `elx-version' did append the "update number" to the
version.

This was done for the benefit of the Emacsmirror but I am no longer
doing that here, and I am doing it differently (by ignoring the
"version" completely, and using the "update number" in its place).