elx-licensee: Deal with licensee errors
Without this we can get an "End of file while parsing JSON" error with
no indication where that is coming from (which is a problem when using
emir).

Do not use `zerop' because the value returned by `call-process' may be
a string.

`licensee' returns zero even in case of some errors.  If the returned
json cannot be parsed, then it probably is not json but a plain error
message.  Show the text so the user can figure out what happened.