elx-licensee: Deal with more licensee weirdness
`licensee' returns with status 1 if it cannot detect a license, which
makes a lot of sense.  However it also returns with 1 if an unhandled
error occurs, which makes less sense and inspired the changes in the
previous commit, one of which has to be reverted to properly handle
the expected failure case.