Properly deal with libraries with problematic suffixes
Emir is forced to treat files with suffixes ".el.in" and ".el.tmpl"
as libraries and does so by binding `load-suffixes'.

The code that is being replaced, was able to deal with that, but as
a result of calling `file-name-sans-extension' twice, misbehaved on
unusual filenames such as "treepy.el-walker-test.el", treating all
of ".el-walker-test" as the second suffix, which ultimately caused
`elx-main-library-2' to signal an error when it found out that that
file "failed" to provide `treepy'.