Use _ in all non-binding entries in the varlist of COND-let forms
This was not done until now because doing so results in a warning
because the macro expansion did not actually leave any variables
unused.  This was fixed in Emacs 30.1 and I've added a backport to
Transient, which fixes the issue for that package and packages that
depend on it.

It doesn't remove the warning when compiling this package, but I have
decided I *always* want to use `_' when appropriate, and that I can
live with users seeing warnings, if they do not bother to update to
the latest Emacs release in a timely manner.  Alternatively they can
add the backport early in their init file.

Without consistently using `_' one would always have to count parens
to be sure whether a member of the varlist does or does not bind a
variable.  It is very easy to make mistake when writing or reading
such forms, if `_' is not consistently used in the non-binding cases.